Adding DDPG with RLAdventure2
